refactor: moved repeated load autocomplete latlon lookup to hooks
